[Bugfix] 'const size_t' (aka 'const unsigned long') is ambiguous on MacOS (#363)
[Bugfix] 'const size_t' (aka 'const unsigned long') is ambiguous on MacOS

+++ src/c_api_common.cc
@@ -24,7 +24,7 @@ DLManagedTensor* CreateTmpDLManagedTensor(const DGLArgValue& arg) {
 
 PackedFunc ConvertNDArrayVectorToPackedFunc(const std::vector<NDArray>& vec) {
     auto body = [vec](DGLArgs args, DGLRetValue* rv) {
-        const size_t which = args[0];
+        const int which = args[0];
         if (which >= vec.size()) {
             LOG(FATAL) << "invalid choice";
         } else {
